- refactored components - Added filter button - dynamic ui with useState hook
