+++ src/components/header/Header.jsx
@@ -1,0 +1,23 @@
+import Foodie from "../../assets/Foodie.png";
+
+const Header = () => {
+  return (
+    <>
+      <div className="header">
+        <div className="logo-container">
+          <img className="logo" src={Foodie} />
+        </div>
+        <div className="nav-items">
+          <ul>
+            <li>Home</li>
+            <li>About Us</li>
+            <li>Contact Us</li>
+            <li>Cart</li>
+          </ul>
+        </div>
+      </div>
+    </>
+  );
+};
+
+export default Header;
